Numerous small polar plot tweaks
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@6494 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/gp/PolarPlot.h
@@ -12,7 +12,7 @@ typedef void *PPCookie;
 extern PPCookie pol_DisplaySetup (int project, int tfill);
 extern PPCookie pol_GridSetup (int, DestImage *, double, double, double,
 		double);
-extern void pol_PlotBeam (PPCookie, PolarBeam *, void *, int, int);
+extern void pol_PlotBeam (PPCookie, PolarBeam *, void *, float, float);
 extern void pol_Finished (PPCookie);	
 
 # ifdef __cplusplus
